Remove query/type from biosampletables file
